setup state code to be ready for backend, moved LoginPage.js to /src, code cleanup in a few places
