+++ mern/react-app/src/App.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import LoginPage from './components/LoginPage/LoginPage'
+import LoginPage from './LoginPage'
 import UniformAssignPage from './UniformAssignPage'
 import ReportsPage from './ReportsPage'
 import { BrowserRouter as Router, Switch, Route } from "react-router-dom";
@@ -17,7 +17,7 @@ function App() {
       </div>
     </Router>
   );
-  
+
 }
 
 export default App;
