+++ mern/react-app/src/UniformAssignPage.module.css
@@ -11,7 +11,6 @@
   width: 100%;
   top: 0px;
   left: 0px;
-  background-color: blue;
 }
 
 .footerWrapper {
@@ -20,7 +19,6 @@
   width: 100%;
   bottom: 0px;
   left: 0px;
-  background-color: blue;
 }
 
 .studentSelectComponentWrapper {
